+++ components/GalleryComponents/Gallery/index.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 import LightBox from "../LightBox"
-import { Cross } from "../icons"
+import { Cross } from "../../icons"
 import { useEffect } from "react"
 import Modal from "@mui/material/Modal"
 import Slide from "@mui/material/Slide"
