+++ components/GalleryComponents/NestedModal/NestedModal.tsx
@@ -4,7 +4,7 @@ import Modal from "@mui/material/Modal"
 import Slide from "@mui/material/Slide"
 
 import Backdrop from "@mui/material/Backdrop"
-import Gallery from "../../components/Gallery"
+import Gallery from "../Gallery"
 import { Image, Row } from "react-bootstrap"
 
 const style = {
