+++ styles/components.scss
@@ -304,6 +304,16 @@
     }
 }
 
+.galleryBox {
+    position: absolute;
+    width: 100%;
+    height: 100vh;
+    border: 2px solid #000;
+    box-shadow: 24;
+    overflow: auto;
+    background-color: black;
+}
+
 // -------------------------
 // Prayer Points
 // -------------------------
